Improve storage lookup performance
Some code I had left from the weekend, where I looked into
the cost of querying the store during eval, and if that could
be improved. The code here is mostly benchmarks that I found
useful for that purpose, but also includes a few optimizations
of code that is on the hot path for these lookups and as such
has a high impact — as the benchmarks demonstrate.

The small change in eval.go is the first use of the new feature
of object.Map added some days ago, where a map function returning
a nil key means skipping that key/value pair.

Signed-off-by: Anders Eknert <anders@styra.com>